pyproject.toml
[project]
name = "repl-nix-pawprintclassifier"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "python-multipart>=0.0.19",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "uvicorn>=0.32.1",
]
